+++ src/Events/ProcessStarted.php
@@ -1,0 +1,84 @@
+<?php
+
+/**
+ * Copyright (c) 2011-present MediaSift Ltd
+ * Copyright (c) 2015-present Ganbaro Digital Ltd
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions
+ * are met:
+ *
+ *   * Redistributions of source code must retain the above copyright
+ *     notice, this list of conditions and the following disclaimer.
+ *
+ *   * Redistributions in binary form must reproduce the above copyright
+ *     notice, this list of conditions and the following disclaimer in
+ *     the documentation and/or other materials provided with the
+ *     distribution.
+ *
+ *   * Neither the names of the copyright holders nor the names of his
+ *     contributors may be used to endorse or promote products derived
+ *     from this software without specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
+ * FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
+ * COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
+ * INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
+ * BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+ * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+ * CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
+ * LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
+ * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+ * POSSIBILITY OF SUCH DAMAGE.
+ *
+ * @category  Libraries
+ * @package   ProcessRunner/Events
+ * @author    Stuart Herbert <stuherbert@ganbarodigital.com>
+ * @copyright 2011-present MediaSift Ltd www.datasift.com
+ * @copyright 2015-present Ganbaro Digital Ltd www.ganbarodigital.com
+ * @license   http://www.opensource.org/licenses/bsd-license.php  BSD License
+ * @link      http://code.ganbarodigital.com/php-process-runner
+ */
+
+namespace GanbaroDigital\ProcessRunner\Events;
+
+use GanbaroDigital\EventStream\Events\Event;
+
+class ProcessStarted implements Event
+{
+	/**
+	 * the command that has been executed
+	 * @var array|Traverseable
+	 */
+	public $command;
+
+	/**
+	 * how long is the command allowed to run, in seconds?
+	 * @var int|null
+	 */
+	public $timeout;
+
+	/**
+	 * which folder do we run the command in?
+	 * @var string|null
+	 */
+	public $cwd;
+
+	/**
+     * @param  array|Traversable $command
+     *         the command to execute
+     * @param  int|null $timeout
+     *         how long before we force the command to close?
+     * @param  string|null $cwd
+     *         the folder to run the command inside
+	 */
+    public function __construct($command, $timeout, $cwd)
+    {
+    	$this->command = $command;
+    	$this->timeout = $timeout;
+    	$this->cwd = $cwd;
+    }
+}